+++ services/cacheEnabler.js
@@ -1,21 +1,26 @@
 
 class CacheEnabler {
     constructor() {
+
+        // long TTL
         this.trigger = true;
         if (process.env.CACHE_TRIGGER_ENABLED=="false" || process.env.CACHE_TRIGGER_ENABLED==false) {
             this.trigger = false;
         }
 
+        // long TTL
         this.subscription = true;
         if (process.env.CACHE_SUBSCRIPTION_ENABLED=="false" || process.env.CACHE_SUBSCRIPTION_ENABLED==false) {
             this.subscription = false;
         }
 
+        //default TTL
         this.project = true;
         if (process.env.CACHE_PROJECT_ENABLED=="false" || process.env.CACHE_PROJECT_ENABLED==false) {
             this.project = false;
         }
 
+        //default TTL
         this.request = true;
         if (process.env.CACHE_REQUEST_ENABLED=="false" || process.env.CACHE_REQUEST_ENABLED==false) {
             this.request = false;
@@ -27,10 +32,10 @@ class CacheEnabler {
         }
 
         
-        // this.project_user = true;
-        // if (process.env.CACHE_PROJECT_USER_ENABLED=="false" || process.env.CACHE_PROJECT_USER_ENABLED==false) {
-        //     this.project_user = false;
-        // }
+        this.project_user = true;
+        if (process.env.CACHE_PROJECT_USER_ENABLED=="false" || process.env.CACHE_PROJECT_USER_ENABLED==false) {
+            this.project_user = false;
+        }
 
         // this.user = true;
         // if (process.env.CACHE_USER_ENABLED=="false" || process.env.CACHE_USER_ENABLED==false) {