+++ services/operatingHoursService.js
@@ -12,7 +12,7 @@ class OperatingHoursService {
 
     // winston.debug('O ---> [ OHS ] -> PROJECT ID ', projectId)
     let q = Project.findOne({_id: projectId, status: 100});
-    if (cacheEnabler.project) {      //cache_virtual_problem
+    if (cacheEnabler.project) { 
       q.cache(cacheUtil.longTTL, "projects:id:"+projectId)  //project_cache
       winston.debug('project cache enabled');
     }